Add proper translation generation
When a new translation is generated, it creates a word of a random size
clamped between half the original size and double the original size.
The characters of the words are uniformly chosen from the alphabet.
Translations that have the same character trice in a row are discarded
and a new translation is created.